+++ src/dataset.jl
@@ -3,7 +3,7 @@ using Base.Iterators: flatten
 
 import RecipesBase
 
-export Dataset, AbstractDataset, minima, maxima
+export Dataset, AbstractDataset, SVector, minima, maxima
 export minmaxima, columns, regularize
 
 abstract type AbstractDataset{D, T} end
@@ -99,7 +99,6 @@ end
 ###########################################################################
 # Concrete implementation
 ###########################################################################
-
 """
     Dataset{D, T} <: AbstractDataset{D,T}
 A dedicated interface for datasets.
@@ -144,8 +143,7 @@ Dataset(s::Dataset) = s
 ###########################################################################
 # Dataset(Vectors of stuff)
 ###########################################################################
-Dataset(s::AbstractVector{T}) where {T<:Number} =
-Dataset{1, T}(reinterpret(SVector{1, T}, s))
+Dataset(s::AbstractVector{T}) where {T} = Dataset(SVector.(s))
 
 function Dataset(v::Vector{<:AbstractArray{T}}) where {T<:Number}
     D = length(v[1])